+++ examples/antd-breadcrumbs/src/main.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import 'antd/dist/reset.css'
+
+import "./index.css";
+import App from "./App";
+
+ReactDOM.createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+  <App />
+</React.StrictMode>
+);
+