+++ examples/antd-breadcrumbs/src/pages/contact/list-view/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+interface ListViewProps {
+    id?: number;
+}
+
+const modulePrefix = 'ListView';
+const ListView: React.FC<ListViewProps> = (props: ListViewProps) => {
+    return (<div className={modulePrefix}>ListView</div>);
+};
+
+export default ListView;