+++ examples/antd-breadcrumbs/src/pages/my-page/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+interface MyPageProps {
+    id?: number;
+}
+
+const modulePrefix = 'MyPage';
+const MyPage: React.FC<MyPageProps> = (props: MyPageProps) => {
+    return (<div className={modulePrefix}>MyPage</div>);
+};
+
+export default MyPage;