+++ examples/antd-breadcrumbs/src/pages/project/clients/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+interface ClientProps {
+    id?: number;
+}
+
+const modulePrefix = 'Client';
+const Client: React.FC<ClientProps> = (props: ClientProps) => {
+    return (<div className={modulePrefix}>Client</div>);
+};
+
+export default Client;